fix(amazonq): use neighbor cells as completion context in Notebook #7086
## Problem
VS Code treats each cell in a notebook as a separate editor. As a
result, when building the left- and right-contexts for the completion
from the current editor, we were limited to just the current cell, which
might be very small and/or reference variables and functions defined in
other cells. That meant that completions never used the context of other
cells when making suggestions, and were often _very_ generic.
https://github.com/aws/aws-toolkit-vscode/issues/7031

## Solution
The `extractContextForCodeWhisperer` function now checks if it is being
called in a cell in a Jupyter notebook. If so, it collects the
surrounding cells to use as context, respecting the maximum context
length. During this process, Markdown cells have each line prefixed with
a language-specific comment character.