fix(amazonq): push customizations on startup (#7297)
## Problem

At the startup of the extension, the customization that a user already
decided previously was not being pushed to flare.
The only time we would push the customization to flare was if the
customization was changed.

Otherwise everything else works as expected.


## Solution

On startup, push the customization to flare (if it already exists)

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>