fix(amazonq): Wrong customization being pushed to LSP on start
Problem

On start of the extension when we called getSelectedCustomization() it would
return the wrong customization. This caused the wrong customization to be
pushed to the LSP.

The reason was due to notifyNewCustomizations not being called before getSelectedCustomization()
was used.

Solution

Ensure the call that uses notifyNewCustomizations() runs before getSelectedCustomization is used.

I also refactored to compartmentalize the relevant code.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>