fix: Call auth setup in the correct order
Problem:

Profiles and Customizations were pushed to the LSP before Auth was initialized (and its bearer token pushed to the LPS),
this caused errors since it does not make sense to push profile/customization if a bearer token does not exist.

Solution:

Ensure the bearer token is pushed first, then follow up with pushing the profile/customization after.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>