fix(amazonq): Profile needing to be selected on server restart (#7316)
See individual commits for isolated changes

## Problem

We were seeing the following errors from the Q Language Server on
startup:
- `Amazon Q Profile is not selected for IDC connection type`
- `Amazon Q service is not signed in`

## Solution

We needed to do 2 solutions, each is a separate commit (see their
message). There were also some minor refactors.

In short:
- The Auth bearer token MUST be sent to the Q LSP before Profile is
sent. We were not doing this and it was causing an error
- When sending the Auth to the Q LSP, the startUrl MUST be included in
the request or else it would fail. We thought we were sending it but
based on the logs prefixed with `UpdateBearerToken` it showed`sso` did
not contain the startUrl

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>