feat(amazonq): Command to clear extension cache
We need a way to clear globalState since it looks like some users are getting
in to a bad corrputed state where things stop working.

Uninstall the extension does not clear state, it is intentionally designed that way by
vscode, as we don't want to wipe states on extension updates.

Solution:

This creates a new command "Amazon Q: Clear extension cache" which clears the cache and
reloads the window.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>