feat(amazonq): Command to clear extension cache (#7335)
## Problem

We need a way to clear `globalState` since it looks like some users are
getting in to a bad corrputed state where things stop working.

Uninstalling the extension does not clear state, it is intentionally
designed that way by vscode, as we don't want to wipe states on
extension updates.

## Solution:

This creates a new command "Amazon Q: Clear extension cache" which
clears the cache and reloads the window.

There is safety modal which pops up right before clearing after the
command is selected.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>