refactor(clearCache): Delete the LSP cache dir
Sometimes users MAY modify the LSP cache dir for reasons like development,
and when we clear all user cache we want to ensure the LSP cache is not
corrupted.

So when the user calls the clear cache command from the command palette it will
also clear the LSP cache too.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>