fix(amazonq): throw if no region profiles are available (#7360)
Duplicate PR of https://github.com/aws/aws-toolkit-vscode/pull/7357 into
Flare branch

## Problem
When ListRegionProfile call throttles for a subset of regions, we
currently do not throw, but instead return the available profiles in the
regions where the call succeeded. However, if that list is empty (no
profiles in that region), we return an empty list. This breaks the UI,
and causes a state that is not recoverable

## Solution
Throw an error in the scenario where availableProfiles is empty. This
triggers a retry state in the UI, making the state recoverable.


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.