fix(amazonq): handle existing Flare connection when migrating SSO connections (#7363)
## Problem
Auth migration to LSP is not handled gracefully when a user downgrades
and upgrades to auth on LSP multiple times, causing users to be logged
out if they upgrade a second time

## Solution
In the auth migration script, call the LSP identity server to check if a
token is available. If the token is available, don't migrate the auth
connection. If no token is available, migrate.

Added unit tests for the case.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.