fix(lambda): checking if folder exists (#6899)
## Problem
In the existing Serverless Land project, we need to check if a folder
with the same name already exists before downloading the code. This
could help avoid potential issues that might arise from overwriting
existing files or folders.

## Solution
The code checks if a folder with the same name already exists, and
prompts the user to choose whether to override it or not. This allows
the user to decide how to handle potential conflicts with existing files
or folders.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Co-authored-by: Vandita Patidar <vanditap@amazon.com>