fix(lambda): Revert update to function upload directory selection wizard (#7624)
fixes #7618

## Problem
Introduced in #7601 is a bug where choosing to upload a Lambda function
from a directory would not allow the user to actually select the
directory. In the directory selection window, the button to select is
disabled. This was an inadvertent change when adding new functionality.

## Solution
Revert the breaking change. In addition to switching `canSelectFolders`
to `true` and `canSelectFiles` to `false`, I have reverted the other
changes in the upload flow that were added in the original PR. This is
because these changes added another flow to the upload modal that
doesn't fit with all the other added functionality, and just increases
the complexity of the code without much benefit. Because this was the
only place that used the `lambdaEdits` array from the `utils`, I removed
that as well.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.