feat(sagemaker): Add Autoshutdown support and Fix connect to capitalized space name (#2156)
## Problem
1. Code Editor and JupyterLab spaces support an auto-shutdown feature
that stops the space after a period of user inactivity. Currently, this
feature does not account for activity when a user connects to their
space through VS Code. We need to extend support for auto-shutdown in
this case.
2. User is unable to connect to a SageMaker Space that contains capital
letters in its name. This is because SSH automatically canonicalizes the
hostname when passing it to the ProxyCommand, which includes converting
it to lowercase and performing DNS lookups.
3. There’s a typo in refreshURL where an extra / is being added.

## Solution
1. Track user activity in VS Code using the `UserActivity` class and by
monitoring terminal input and output. Write the latest activity
timestamp to the space's local `/tmp` file so the backend auto-shutdown
system can detect recent user activity.
2. Use the %n placeholder instead of %h in the ProxyCommand. Unlike %h,
which provides the resolved and lowercased hostname, %n retains the
original, user-provided value including capital letters.
3. There’s a typo in refreshURL where an extra / is being added.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.