feat(sagemaker): Show notification if instanceType has insufficient memory (#2157)
## Problem
When the user has not started the space before (instanceType not
defined), the extension would fall back to using `ml.t3.medium`, which
has insufficient memory.
Or if the user has attempted to restart the space with an instanceType
with insufficient memory, the request to `StartSpace` would fail.

Instance types with insufficient memory:
- `ml.t3.medium`
- `ml.c7i.large`
- `ml.c6i.large`
- `ml.c6id.large`
- `ml.c5.large`

## Solution
Show an error notification if user is attempting to start a space with
insufficient memory. Suggest the user to use an upgraded instance type
with more memory depending on the one they attempted to use. If the user
confirms, then the call to `StartSpace` will continue with the
recommended instanceType.

- `ml.t3.medium` --> `ml.t3.large`
- `ml.c7i.large` --> `ml.c7i.xlarge`
- `ml.c6i.large` --> `ml.c6i.xlarge`
- `ml.c6id.large` --> `ml.c6id.xlarge`
- `ml.c5.large` --> `ml.c5.xlarge`

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Co-authored-by: Newton Der <dernewtz@amazon.com>