fix(sagemaker): Show error message when trying to connect remotely from remote workspace (#2158)
## Problem
When a user is connected to a remote workspace and clicks the Open
Remote Connection button, they see an error message about the Remote SSH
extension not being installed, when the real reason is that they cannot
connect remotely from a remote workspace.

## Solution
Show an error message which states clearly that they cannot connect via
deeplink when in a remote workspace.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Co-authored-by: Newton Der <dernewtz@amazon.com>