fix(amazonq): fix Inline completion acceptance and reject telemetry race condition (#7734)
## Problem

When Q is editing (on accept, on reject is in progress), if we trigger
again, the session is not closed yet, global state varaibles are still
in progress to be cleared, and we will report wrong user trigger
decision telemetry. This is worse for acceptance since it has a await
sleep for diagnostics to update.

## Solution

Do not let it trigger when Q is editing!  

This is not customer facing so no change log. 

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.