fix(amazonq): early stop pagination requests when user decision is made (#7759)
## Problem

When a paginated response is in flight, but the user already accepted or
rejected a completion, the subsequent paginated requests should be
cancelled.

This PR also fixes the `codewhispererTotalShownTime` being negative
issue by using performance.now() across all timestamp computation.


## Solution

This is not a user facing change.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.