telemetry(amazonq): flare is now source of truth for metrics (#7768)
## Problem
In VSC, we check that the metric name must be in
[aws-toolkit-common](https://github.com/aws/aws-toolkit-common) before
emitting the metric.

Therefore when we want to add a new metric, the current process is:
1. Add new metric in aws-toolkit-common
2. Wait for version to increment (~1 hour)
3. Bump up toolkit-common version in VSC repo
4. Wait for next VSC release (up to 1 week)

Only after steps 1-4, will we be actually emitting the new metric.

JB, VS, and Eclipse do not have this dependency, and assume Flare is the
source of truth for metrics

## Solution
In VSC, Flare is now the source of truth for metrics instead of
depending on aws-toolkit-common




---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.