feat(appcomposer): Update CFN snippet source (#7729)
## Problem

The AppComposer generate feature is moving from a call to the Q
extension to a call to our owned CDN, the same CDN that hosts our
webview, or CFN template partials.


## Solution

This PR updates the code that used to make a call to the Q extension and
instead replaces it with fetch call to the CDN to gather data. Error
cases (such as the requested file not being found or the CDN being down)
are handled by the existing code.


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.