feat(smus): Add credential providers for SMUS (#2187)
**Description**

Added credential providers for DER, Project role credentials and
Connections credentials. Added in memory caching for now.

The credential providers currently are structured so that it can be used
with CredentialStore but a lot of things to do with credential store is
really old. In the coming days, while we do bug bashes and tests, if the
in memory caching fares well enough, will simplify the credential
provider to implement the AWS SDK CredentialProvider directly.


**Motivation**

Support auth for SMUS and ensure all clients get credentials and don't
fail due to lack of credentials.

**Testing Done**

Updated unit tests and tested flow locally. 

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Co-authored-by: Bhargava Varadharajan <vabharga@amazon.com>