feat(amazonq): add shortcut test for UI E2E Tests (#7863)
## Problem
We have not implemented the shortcut tests for our UI E2E Test suite
with VSCode-Extension-Tester. Additionally, we face a small problem in
our pressShortcut abstraction that throws a type error when trying to
reference a Key command like `Key.COMMAND` since it is a IKey String.

## Solution
We have implemented the shortcut tests: 
- Optimize Code Shortcut 
- Generate Test Shortcut
- Keybind Check
- Explain Code Shortcut

We have also fixed the pressShortcut abstraction with a small edit in
the logic of accepted parameters.

NOTE: We have chosen to implement the Inline Shortcut within the Inline
Tests even though it could be classified under Shorcuts.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.