fix(sagemaker): Reduce number of reconnect retries (#7827)
## Problem
When a connection to a remote space is severed, the extension will
attempt to reconnect up to 60 times. Each reconnect attempt opens the
Space Details page in a new browser tab. Opening up to 60 new tabs is a
poor user experience.

## Solution
We change the maximum number of retries to 8.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Co-authored-by: Newton Der <dernewtz@amazon.com>