fix(explorer): Refresh s3 node on profile switch (#7401, #7665)
When a connection is changed, the awsClientBuilderV3 is used to get the service details, and a cache is being used to cache the service. However, in the cache key only region is cached along with other service details, so on profile switch it is still returning the same cached service of previous profile.

Retrieved current profile using global aws context, and added it to the cache key, so when profile is switched, cache is updated and service from new profile is retrieved.