fix(amazonq): XPath Selector for pinContextHelpers and fixing PinContext ability to run in full test suite (#7928)
## Problem
We were faced with the problem that when all our tests and helpers were
merged together, and then run as a full suite, we saw errors with the
pinContext.test.ts file. Namely, it was targeting a background overlay
that had identical css components to the pinContext overlay. This is
problematic since our entire suite relies on these css components, and
even test-ids were not differentiated. Additionally, our
pinContextHelpers used sleep functions in the range of 100-3000 ms which
is naturally arbitrary values.

## Solution
We implement the XPath css selectors that are used within Selenium to
identify the pinContext Buttons by text, not css classes or ids. This
drastically decreases the length and complexity of our helpers while
also eliminating the need for sleep functions other than a sleep(0)
function for subMenuItems that is needed for the DOM to be loaded. Below
is our success of the tests:

<img width="313" height="77" alt="Screenshot 2025-08-19 at 1 49 34 PM"
src="https://github.com/user-attachments/assets/dd378c02-67c1-4f73-a6e7-cd709e028c80"
/>

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.