fix(smus): Add proactive cred refresh for active SSH connections
**Description**

Added a proactive cred check and refresh when SSH connections
are established.

Also updated the error messages to be actionable for users.

**Motivation**

Bug : Previously once cred expired, we were throwing blanket
error which did not tell user what the issue was and there
was no path to recovery as well. Now with proactive cred refresh,
user should be able to retry in ~10-15 seconds.

**Testing Done**

Tested all flows manually. Unit tests partial, needs to be updated.