fix(toolkit): correctly use the endpoint URL for the selected profile
https://github.com/aws/aws-toolkit-vscode/issues/2007

The Toolkit doesn't read the endpoint configured in an AWS profile
config.

The endpoint even before these changes is already being read from the
AWS config file, and it's included in the CredentialsProvider object,
but there wasn't a way to retrieve that. We're adding `getEndpointUrl()`
method to the different providers, and then we're retrieving that
endpoint in:
1. when loading the current profile, to add it to the credentialsCache
(`auth.ts -> createCachedCredentials()`, `loginManager.ts -> login()`).
Because of this, then we can get this endpoint alongside the credentials
when creating a new AWS client.
2. when creating the list of connections that then will be displayed for
the user to pick (`packages/core/src/auth/auth.ts ->
getIamConnection()`) (this required to make `getIamConnection` an async