feat(smus): Support IAM auth mode for SMUS (#2236)
**Description**
Adding the UI flow to support IAM auth flow in SMUS. The text is TBD and
is placeholder for now and the actual core IAM flow and tracking is not
implemented yet. Will be in the next CR.

**Motivation**

**Testing Done**
Added unit tests where possible and also manually tested existing and
new UI flow.

## Problem


## Solution


---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Co-authored-by: Bhargava Varadharajan <vabharga@amazon.com>