feat(sagemakerunifiedstudio): Support remote access for Kiro IDE (#2242)
## Problem

SageMaker AI Studio and SageMaker Unified Studio support the ability for
customers to connect their local VS Code IDE to SageMaker Spaces from
the AWS Toolkit extension, but this is not supported for the Kiro IDE.
The primary reason is that the VS Code Remote SSH extension is not
available for other IDEs, and an SSH extension is an essential component
of the remote access feature architecture.

## Solution

This PR adds a new `SageMaker Remote SSH` extension to the AWS Toolkit
VS Code repository, forked from
https://github.com/jeanp413/open-remote-ssh, which performs a similar
function to the VS Code Remote SSH extension. It doesn't have as many
features as the VS Code version (which is closed source), but it
suffices for simple use cases.

Several notes about this implementation:
- Overall, this additional extension is intended to be a temporary
solution (e.g., no longer than six months), as we are expecting the Kiro
team to eventually provide a built-in SSH solution that is more feature
rich (i.e., supports other types of hosts, not just SageMaker Spaces),
though no timeline has been defined.
- The `SageMaker Remote SSH` extension name, description and icon are
not finalized, currently pending internal review. For now, it is using
the AWS Toolkit icon, which is just saved directly into the repository
to simplify the build process.
- The reason that we have created a separate extension, as opposed to
simply authoring the code directly into the AWS Toolkit extension, is
that SSH extensions must be configured as `ui` extensions, whereas the
AWS Toolkit is a `workspace` extension (see
https://code.visualstudio.com/api/advanced-topics/extension-host).
- The SageMaker Remote SSH extension code is authored as a workspace
under `packages/sagemaker-remote-ssh`. This means it shares
`node_modules` with the rest of the repository. This simplifies some
things such as generating attribution notices. I originally attempted
another approach, where the package is nested somewhere under
`packages/toolkit`, but I faced some typescript errors with duplicate
declaration files being detected.
- The SageMaker Remote SSH extension VSIX file is not intended to be
distributed separately, but rather is packaged under the `toolkit`
extension's `resources` directory as part of the build process.
- When a user attempts to connect to a SageMaker Space from within the
Kiro IDE, the user will be prompted to install SageMaker Remote SSH
extension from disk from the AWS Toolkit extension's `resources`
directory.
- VS Code users should not be affected by this change, ensured by
conditionally checking the IDE name at the relevant places.
- The Deep Link feature of SageMaker AI Studio, which allows customers
to ingress into the IDE from the SageMaker Studio web portal, is not
being updated for the Kiro IDE at this time. The SageMaker Unified
Studio doesn't support this feature today even for VS Code. Either the
SMAI or SMUS team may follow-up on this at a later date.
- The `open-remote-ssh` repository was originally licensed under MIT,
but we are re-licensing it under Apache-2.0. All source files from the
original extension have a header which credits the author and preserves
the original copyright notice.
- In order to support the VS Code proposed `resolvers` API (to function
as an SSH extension), we used the `vscode-dts` utility to generate the
`vscode.proposed.resolvers.d.ts` file, which we have simply committed to
the codebase for simplicity, as we don't expect to need to update it
later.
- Most of the code from the original `open-remote-ssh` extension has
been deleted, in order to make it mostly invisible to users (i.e., no
user interface for browsing hosts, no registered commands), and to
prevent conflicts with that extension if the user has it installed.
- The `open-remote-ssh` extension did not come with any unit tests, and
we haven't added any for the forked version (`sagemaker-remote-ssh`).
There is room for pushback here, but we do not want to invest in
comprehensively unit-testing essentially 3rd party code.
- ~~The SageMaker Remote SSH extension's TypeScript configuration has
been relaxed, and there are numerous ESLint overrides, since most of the
code came from `open-remote-ssh` and we don't want to invest in updating
it all.~~

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.