fix(sagemaker): Adjust retry configuration for StartSession (#8219)
## Problem
When a user tries to connect to a SageMaker Space that is in the Stopped
status (i.e., the underlying App is Deleted or has not been created),
the Space will be automatically started by the toolkit before the
connection is attempted. In some cases, the Space reaches the Running
status (i.e., the App reaches the InService status) but the remote
access capability is not yet ready as it starts asynchronously, leading
to the SageMaker:StartSession API receiving an Internal Failure
response. The client already retries, but the retries happen too
quickly, before remote access becomes ready.

## Solution
Adjust the SageMaker client retry configuration for StartSession calls
made from the detached server (called via the `sagemaker_connect`
script) to spread out the retries over multiple seconds.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Co-authored-by: Laxman Reddy <141967714+laileni-aws@users.noreply.github.com>